Refactor and cleanup [ci fast]
