+++ modules/nextflow/src/main/groovy/nextflow/cli/CmdCid.groovy
@@ -29,6 +29,7 @@ import nextflow.plugin.Plugins
 import org.pf4j.ExtensionPoint
 
 /**
+ * CID command line interface
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
@@ -38,7 +39,7 @@ class CmdCid extends CmdBase implements UsageAware {
 
     private static final String NAME = 'cid'
 
-    interface CidOperation extends ExtensionPoint {
+    interface CidCommand extends ExtensionPoint {
         void log(ConfigMap config)
         void show(ConfigMap config, List<String> args)
         void lineage(ConfigMap config, List<String> args)
@@ -53,7 +54,7 @@ class CmdCid extends CmdBase implements UsageAware {
 
     private List<SubCmd> commands = new ArrayList<>()
 
-    private CidOperation operations
+    private CidCommand operation
 
     private ConfigMap config
 
@@ -86,8 +87,8 @@ class CmdCid extends CmdBase implements UsageAware {
         // init plugins
         Plugins.load(config)
         // load the command operations
-        this.operations = Plugins.getExtension(CidOperation)
-        if( !operations )
+        this.operation = Plugins.getExtension(CidCommand)
+        if( !operation )
             throw new IllegalStateException("Unable to load CID plugin")
         // consume the first argument
         getCmd(args).apply(args.drop(1))
@@ -161,7 +162,7 @@ class CmdCid extends CmdBase implements UsageAware {
                 usage()
                 return
             }
-            operations.log(config)
+            operation.log(config)
         }
 
         @Override
@@ -190,7 +191,7 @@ class CmdCid extends CmdBase implements UsageAware {
                 return
             }
 
-            operations.show(config, args)
+            operation.show(config, args)
         }
 
         @Override
@@ -217,7 +218,7 @@ class CmdCid extends CmdBase implements UsageAware {
                 return
             }
 
-            operations.lineage(config, args)
+            operation.lineage(config, args)
         }
 
         @Override
